+++ src/middleware.ts
@@ -2,7 +2,6 @@ import { NextRequest, NextResponse } from "next/server";
 import { auth } from "./auth";
 
 export default async function middleware(request: NextRequest) {
-  console.log("미들웨어 호출", request.nextUrl.href);
   const session = await auth();
 
   if (!session?.user) {
